+++ src/utils/commands.ts
@@ -136,3 +136,8 @@ export async function moveTab(
     }
   }
 }
+
+export async function moveTabToEnd(viewColumn: number, index: number): Promise<void> {
+  await focusTabInGroup(viewColumn, index);
+  await commands.executeCommand('workbench.action.moveEditorToEnd');
+}